% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bug_report.R
\name{mod_bug_report_ui}
\alias{mod_bug_report_ui}
\alias{mod_bug_report_server}
\title{mod_bug_report_ui and mod_bug_report_server}
\usage{
mod_bug_report_ui(id)

mod_bug_report_server(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
