% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_build_design_example.R
\name{mod_build_design_example_ui}
\alias{mod_build_design_example_ui}
\alias{mod_build_design_example_server}
\title{mod_build_design_example_ui and mod_build_design_example_server}
\usage{
mod_build_design_example_ui(id)

mod_build_design_example_server(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{designLevel}{An integer}
}
\description{
A shiny Module.
}
\keyword{internal}
