% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_check_updates.R
\name{mod_check_updates_ui}
\alias{mod_check_updates_ui}
\alias{mod_check_updates_server}
\title{mod_check_updates_ui and mod_check_updates_server}
\usage{
mod_check_updates_ui(id)

mod_check_updates_server(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\value{
xxxx
}
\description{
A shiny Module.
}
\keyword{internal}
