% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_choose_pipeline.R
\name{mod_choose_pipeline_ui}
\alias{mod_choose_pipeline_ui}
\alias{mod_choose_pipeline_server}
\title{mod_choose_pipeline_ui and mod_choose_pipeline_server}
\usage{
mod_choose_pipeline_ui(id)

mod_choose_pipeline_server(id, package = NULL)
}
\arguments{
\item{id}{shiny id}

\item{package}{xxx}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{dataType}{xxx}
}
\description{
A shiny Module.
}
\keyword{internal}
