% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_format_DT.R
\name{mod_format_DT_ui}
\alias{mod_format_DT_ui}
\alias{mod_format_DT_server}
\title{mod_format_DT_ui and mod_format_DT_server}
\usage{
mod_format_DT_ui(id)

mod_format_DT_server(
  id,
  table2show,
  withBtns = NULL,
  showRownames = FALSE,
  dom = "Bt",
  style = NULL
)
}
\arguments{
\item{id}{shiny id}

\item{withBtns}{xxx}

\item{showRownames}{xxxx}

\item{dom}{xxx}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
