% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_homepage.R
\name{mod_homepage_ui}
\alias{mod_homepage_ui}
\alias{mod_homepage_server}
\title{mod_homepage_ui and mod_homepage_server}
\usage{
mod_homepage_ui(id)

mod_homepage_server(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
