% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_insert_md.R
\name{mod_insert_md_ui}
\alias{mod_insert_md_ui}
\alias{mod_insert_md_server}
\title{mod_insert_md_ui and mod_insert_md_server}
\usage{
mod_insert_md_ui(id)

mod_insert_md_server(id, url)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
