% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loading_page.R
\name{mod_loading_page_ui}
\alias{mod_loading_page_ui}
\alias{mod_loading_page_server}
\title{mod_loading_page_ui and mod_loading_page_server}
\usage{
mod_loading_page_ui(id)

mod_loading_page_server(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
