% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_navbar_menu.R
\name{mod_navbar_menu_ui}
\alias{mod_navbar_menu_ui}
\alias{mod_navbar_menu_server}
\title{mod_navbar_menu_ui and mod_navbar_menu_server}
\usage{
mod_navbar_menu_ui(id)

mod_navbar_menu_server(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
