% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_open_demoDataset.R
\name{mod_open_demoDataset_ui}
\alias{mod_open_demoDataset_ui}
\alias{mod_open_demoDataset_server}
\title{mod_open_demo_dataset_ui and mod_open_demo_dataset_server}
\usage{
mod_open_demoDataset_ui(id)

mod_open_demoDataset_server(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{pipeline.def}{xxx}
}
\value{
An object of class [`xxxx`]
}
\description{
A shiny Module.
}
\keyword{internal}
