% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popover_for_help.R
\name{mod_popover_for_help_ui}
\alias{mod_popover_for_help_ui}
\alias{mod_popover_for_help_server}
\title{mod_popover_for_help_ui and mod_popover_for_help_server}
\usage{
mod_popover_for_help_ui(id)

mod_popover_for_help_server(id, data)
}
\arguments{
\item{id}{shiny id}

\item{data}{A list of two items:
* title: xxxx
* content: xxxx}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
