% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settings.R
\name{mod_settings_ui}
\alias{mod_settings_ui}
\alias{mod_settings_server}
\title{mod_settings_ui and mod_settings_server}
\usage{
mod_settings_ui(id)

mod_settings_server(id, obj)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
